// src/user/payments/dto/payment.enums.ts
// src/payments/payment.enums.ts

export enum PaymentStatus {
  READY = 'READY',
  IN_PROGRESS = 'IN_PROGRESS',
  WAITING_FOR_DEPOSIT = 'WAITING_FOR_DEPOSIT',
  DONE = 'DONE',
  CANCELED = 'CANCELED',
  ABORTED = 'ABORTED',
  FAILED = 'FAILED',
  PENDING = 'PENDING',
}

export enum PaymentMethod {
  CARD = 'CARD',
  VIRTUAL_ACCOUNT = 'VIRTUAL_ACCOUNT',
  TRANSFER = 'TRANSFER',
  MOBILE_PHONE = 'MOBILE_PHONE',
  GIFT_CERTIFICATE = 'GIFT_CERTIFICATE',
  CASH_RECEIPT = 'CASH_RECEIPT',
  EASY_PAY = '간편결제',
}

export enum ReservationStatus {
  PENDING = 'PENDING',
  CONFIRMED = 'CONFIRMED',
  CANCELED = 'CANCELED',
  COMPLETED = 'COMPLETED',
  IN_PROGRESS = 'IN_PROGRESS',
  REJECTED = 'REJECTED',
  NO_SHOW = 'NO_SHOW',
  BLOCKED = 'BLOCKED',
}